testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/location_search.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>location_search</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		if (('location' in window) && window.location != null
			&& ('location' in document) && document.location != null){
			window.history.pushState(1, document.title, '?x=1');
			var search = location.search;

			test(function() {
				assert_equals(search, "?x=1", "search attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("location interface is not supported.");
			});
		}

	</script>
 </body>
</html>
